requirements.txt
fastapi==0.65.2
mysql-connector==2.2.9
protobuf==3.15.8
pydantic==1.8.2
six==1.15.0
starlette==0.14.2
typing-extensions==3.10.0.0
